pyproject.toml
[project]
name = "skinbit"
version = "0.1.0"
description = "Instance-level image matching for dermoscopic images"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "timm>=0.9.0",
    "hydra-core>=1.3.0",
    "wandb>=0.15.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.10.0",
    "pillow>=10.0.0",
    "tqdm>=4.65.0",
    "omegaconf>=2.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]
